requirements.txt
# Core AI & Graph
anthropic>=0.49.0
neo4j>=5.18.0
python-dotenv>=1.0.0

# Embeddings (for Layer 2 typology RAG)
openai>=1.30.0

# Data loading
pandas>=2.2.0

# Layer 2 — typology document pipeline
pdfplumber>=0.11.0
pyyaml>=6.0.0

# App
streamlit>=1.35.0
plotly>=5.22.0

# Notebooks
jupyter>=1.1.0
jupyterlab>=4.0.0
ipykernel>=6.29.0
ipywidgets>=8.0.0

# MCP
fastmcp>=0.1.0
mcp-neo4j-cypher
mcp

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0
